fix: Change LibraryProvider to use React Context, not store.
Using the store caused some rendering issues when the sidebar opened
or closed. Those issues are now resolved, although our app setup
code is now locked into the world of React (and not portable, like our
Svelte store code ironically is).

Is locking myself into the world of React something I want to do?
Sure, right now it is. Maybe I'll want to change that later. I hope that
the fact that I could do this change in less than an hour is evidence
that it is _possible_ to replace these systems — although I know for sure
that as Citadel grows more complex, a "quick and dirty" change like this
won't be possible.