Support updating & inserting book identifiers
Boy, am I getting tired of this domain-driven-design crap. `libcalibre` is too small for it too matter. If I split up the library into entities, a user-facing interface, and an internal persistence-layer, that would be fine. ...honestly, I think that's what DDD is, but with more layers. Probably merging the book, author, and library service into one would be enough to make the library easier to write.